requirements.txt
backcall==0.2.0
Brotli==1.0.9
cchardet==2.1.7
click==8.0.3
colorama==0.4.4
dash==2.0.0
dash-bootstrap-components==1.0.0rc1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
detect-delimiter==0.1.1
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
plotly==5.3.1
prompt-toolkit==3.0.20
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pyzmq==22.3.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
Werkzeug==2.0.2
